refactor: streamline ExpenseScreen and IncomeScreen with new features
- Removed the IncomeScreenRefactored and integrated its functionality into the IncomeScreen for a more cohesive structure.
- Enhanced ExpenseScreen with a new 'Add Expense' button and improved layout for better user interaction.
- Implemented a date filtering feature in both screens, allowing users to filter transactions by date range.
- Added a new empty state representation in the IncomeScreen for better user feedback when no income is recorded.
- Refactored transaction grouping logic to categorize transactions based on date, improving data organization and visibility.
